refactor tests so both the api and CLI tests use sharding without duplicate code
